webapp/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Werkzeug==2.3.7
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
python-dotenv==1.0.0
Flask-WTF==1.1.1
email-validator==2.0.0.post2
plotly==5.15.0
dash==2.11.1
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
openai>=1.0.0
